refactor(calendar-web): rename utils to style-utils
